requirements.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
Django==5.2.8
django-celery-beat==2.8.1
django-timezone-field==7.1
djangorestframework==3.16.1
filelock==3.20.0
fsspec==2025.10.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
kombu==5.5.4
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
nltk==3.9.2
numpy==2.3.4
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
pyphen==0.17.2
python-crontab==3.3.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
redis==7.0.1
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
textstat==0.7.11
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
